Some various fixes

See merge request editions/research-development/glpi-greenit/carbon!33

+++ src/CarbonEmission.php
@@ -100,13 +100,14 @@ class CarbonEmission extends CommonDBChild
         return $input;
     }
 
-    /** Format a weight passing a weight in grams
+    /**
+     * Format a weight passing a weight in grams
      *
-     * @param integer $weight  Weight in grams
+     * @param float $weight  Weight in grams
      *
      * @return string  formatted weight
      **/
-    public static function getWeight(int $weight): string
+    public static function getWeight(float $weight): string
     {
        //TRANS: list of unit (o for octet)
         $units = [
